Factored out dag_log_recover.
