+++ makeflow/src/Makefile
@@ -6,7 +6,7 @@ include ../../rules.mk
 LOCAL_LINKAGE=$(CCTOOLS_GLOBUS_LDFLAGS)
 
 EXTERNAL_DEPENDENCIES = ../../work_queue/src/libwork_queue.a ../../chirp/src/libchirp.a ../../dttools/src/libdttools.a
-OBJECTS = dag.o visitors.o makeflow_common.o lexer.o parser.o makeflow_summary.o
+OBJECTS = dag.o dag_log.o visitors.o makeflow_common.o lexer.o parser.o makeflow_summary.o
 PROGRAMS = makeflow makeflow_viz makeflow_analyze makeflow_linker
 SCRIPTS = condor_submit_makeflow makeflow_graph_log makeflow_monitor starch makeflow_linker_perl_driver makeflow_linker_python_driver
 TARGETS = $(PROGRAMS)
