+++ makeflow/src/dag_log.h
@@ -1,0 +1,10 @@
+
+#ifndef DAG_LOG_H
+#define DAG_LOG_H
+
+#include "dag.h"
+
+void dag_log_recover( struct dag *d, const char *filename, int verbose_mode );
+
+#endif
+
